Set ra_plugin when setting up replica CA
With ra_plugin unconfigured the other CA backends are not
available during installation. The methods that utilize this,
like those that use the dogtag backend directly, were
purposefully skipped, I guess for this reason.

The effect of not enabling ra_plugin is that a lot of
backends are missing from api.Backend including kra,
ra, ra_certprofile and ra_lightweight_ca.

Also force enable_ra to False in this API initialization
to ensure that no communciation to the local CA is
attempted.

https://pagure.io/freeipa/issue/8738

Signed-off-by: Rob Crittenden <rcritten@redhat.com>