Add a --quiet option for add and mod operations
This skips the post_callback step, saving an LDAP query to
retrieve the added/updated entry, massaging the output and
returning it to the caller. A completely empty result is
returned and on the CLI nothing will be shown.

There are some plugins where work is done in the post_callback
step. For example in user the user is added to the default
group. For cases like this where post_callback cannot be
skipped a new CRUD class option was added, no_quiet. If
no_quiet is True then the quiet option isn't added as an
option.

In some plugins this can save multiple LDAP queries. User and
host are examples as they do presence searches on userpassword
and krbprincipalkey.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>