ipatests: test for allowing duplicate cert subjects in ipa-cacert-manage
Use a pre-generated external CA plus two subordinate CA certificates
that use the same subject but have different serial numbers and a
validity period 3years+1day different.

Test that both certificates can be loaded and applied on a system.

If the duplicate subject certificates are not sufficiently
different in validity period, or prior to the this fix,
the test will fail because only one of the duplicately named
subject certificates will be visible: the second one (4097).

Fixes: https://pagure.io/freeipa/issue/9652

Signed-off-by: Rob Crittenden <rcritten@redhat.com>